[ASAP-1662] - Relocate compliance dashboard to projects area
Moves Compliance.tsx and its test from network/teams to projects, keeping
the data layer where it is and importing it cross-directory, the same shape
ProjectWorkspace already uses.

The teams data-layer imports are fully qualified because src/projects has
its own api.ts, export.ts and state.ts — a bare ./api would resolve to the
wrong module.

TeamProfile's lazy loader is re-pointed at the new location so the team tab
keeps working until it is removed.
